Refactored common code across go servers
