+++ go/internal/feast/onlinestore/cassandraonlinestore.go
@@ -2,25 +2,23 @@ package onlinestore
 
 import (
 	"context"
-	"encoding/binary"
 	"encoding/hex"
 	"errors"
 	"fmt"
+	"os"
+	"strings"
+	"time"
+
 	"github.com/feast-dev/feast/go/internal/feast/registry"
+	"github.com/feast-dev/feast/go/internal/feast/utils"
 	"github.com/feast-dev/feast/go/protos/feast/serving"
 	"github.com/feast-dev/feast/go/protos/feast/types"
 	"github.com/gocql/gocql"
 	"github.com/golang/protobuf/proto"
 	"github.com/rs/zerolog/log"
+
 	"google.golang.org/protobuf/types/known/timestamppb"
-	_ "google.golang.org/protobuf/types/known/timestamppb"
 	gocqltrace "gopkg.in/DataDog/dd-trace-go.v1/contrib/gocql/gocql"
-	_ "net"
-	"os"
-	"sort"
-	"strings"
-	"time"
-	_ "time"
 )
 
 type CassandraOnlineStore struct {
@@ -50,8 +48,8 @@ func NewCassandraOnlineStore(project string, config *registry.RepoConfig, online
 	// Parse host_name and Ips
 	cassandraHosts, ok := onlineStoreConfig["hosts"]
 	if !ok {
-		cassandraHosts = "127.0.0.1"
-		log.Warn().Msg("Host not provided: Using localhost instead")
+		cassandraHosts = []interface{}{"127.0.0.1"}
+		log.Warn().Msg("host not provided: Using 127.0.0.1 instead")
 	}
 
 	var rawCassandraHosts []interface{}
@@ -65,19 +63,19 @@ func NewCassandraOnlineStore(project string, config *registry.RepoConfig, online
 		if !ok {
 			return nil, fmt.Errorf("failed to convert a host to a string: %+v", rawHost)
 		}
-		fmt.Printf("\tsingle host: %s\n", hostStr)
 		cassandraHostsStr[i] = hostStr
 	}
 
 	username, ok := onlineStoreConfig["username"]
 	if !ok {
-		username = "scylla"
-		log.Warn().Msg("Username not defined: Using default username instead")
+		username = "cassandra"
+		log.Warn().Msg("username not defined: Using default username instead")
 	}
+
 	password, ok := onlineStoreConfig["password"]
 	if !ok {
-		password = "scylla"
-		log.Warn().Msg("Password not defined: Using default password instead")
+		password = "cassandra"
+		log.Warn().Msg("password not defined: Using default password instead")
 	}
 
 	var usernameStr string
@@ -143,7 +141,7 @@ func NewCassandraOnlineStore(project string, config *registry.RepoConfig, online
 	store.clusterConfigs.Authenticator = gocql.PasswordAuthenticator{Username: usernameStr, Password: passwordStr}
 	createdSession, err := store.clusterConfigs.CreateSession()
 	if err != nil {
-		return nil, fmt.Errorf("Unable to connect to the ScyllaDB database")
+		return nil, fmt.Errorf("unable to connect to the ScyllaDB database")
 	}
 	store.session = createdSession
 	return &store, nil
@@ -152,6 +150,7 @@ func NewCassandraOnlineStore(project string, config *registry.RepoConfig, online
 func (c *CassandraOnlineStore) getFqTableName(tableName string) string {
 	return fmt.Sprintf(`"%s"."%s_%s"`, c.keyspace, c.project, tableName)
 }
+
 func (c *CassandraOnlineStore) getCQLStatement(tableName string, featureNames []string, nKeys int) string {
 	// TODO: Compare with multiple single-key concurrent queries like in the Python feature server
 	keyPlaceholders := make([]string, nKeys)
@@ -177,7 +176,7 @@ func (c *CassandraOnlineStore) buildCassandraEntityKeys(entityKeys []*types.Enti
 	cassandraKeys := make([]interface{}, len(entityKeys))
 	cassandraKeyToEntityIndex := make(map[string]int)
 	for i := 0; i < len(entityKeys); i++ {
-		var key, err = serializeCassandraEntityKey(entityKeys[i], c.config.EntityKeySerializationVersion)
+		var key, err = utils.SerializeEntityKey(entityKeys[i], c.config.EntityKeySerializationVersion)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -200,7 +199,7 @@ func (c *CassandraOnlineStore) OnlineRead(ctx context.Context, entityKeys []*typ
 	serializedEntityKeys, serializedEntityKeyToIndex, err := c.buildCassandraEntityKeys(entityKeys)
 
 	if err != nil {
-		return nil, fmt.Errorf("Error when serializing entity keys for Cassandra")
+		return nil, fmt.Errorf("error when serializing entity keys for Cassandra")
 	}
 	results := make([][]FeatureData, len(entityKeys))
 	for i := range results {
@@ -293,96 +292,6 @@ func (c *CassandraOnlineStore) OnlineRead(ctx context.Context, entityKeys []*typ
 
 	return results, nil
 
-}
-
-// Serialize entity key to a bytestring so that it can be used as a lookup key in a hash table.
-func serializeCassandraEntityKey(entityKey *types.EntityKey, entityKeySerializationVersion int64) (*[]byte, error) {
-	// Ensure that we have the right amount of join keys and entity values
-	if len(entityKey.JoinKeys) != len(entityKey.EntityValues) {
-		return nil, fmt.Errorf("the amount of join key names and entity values don't match: %s vs %s", entityKey.JoinKeys, entityKey.EntityValues)
-	}
-	// Make sure that join keys are sorted so that we have consistent key building
-	m := make(map[string]*types.Value)
-
-	for i := 0; i < len(entityKey.JoinKeys); i++ {
-		m[entityKey.JoinKeys[i]] = entityKey.EntityValues[i]
-	}
-
-	keys := make([]string, 0, len(m))
-	for k := range entityKey.JoinKeys {
-		keys = append(keys, entityKey.JoinKeys[k])
-	}
-	sort.Strings(keys)
-
-	// Build the key
-	length := 5 * len(keys)
-	bufferList := make([][]byte, length)
-
-	for i := 0; i < len(keys); i++ {
-		offset := i * 2
-		byteBuffer := make([]byte, 4)
-		binary.LittleEndian.PutUint32(byteBuffer, uint32(types.ValueType_Enum_value["STRING"]))
-		bufferList[offset] = byteBuffer
-		bufferList[offset+1] = []byte(keys[i])
-	}
-
-	for i := 0; i < len(keys); i++ {
-		offset := (2 * len(keys)) + (i * 3)
-		value := m[keys[i]].GetVal()
-
-		valueBytes, valueTypeBytes, err := serializeCassandraValue(value, entityKeySerializationVersion)
-		if err != nil {
-			return valueBytes, err
-		}
-
-		typeBuffer := make([]byte, 4)
-		binary.LittleEndian.PutUint32(typeBuffer, uint32(valueTypeBytes))
-
-		lenBuffer := make([]byte, 4)
-		binary.LittleEndian.PutUint32(lenBuffer, uint32(len(*valueBytes)))
-
-		bufferList[offset+0] = typeBuffer
-		bufferList[offset+1] = lenBuffer
-		bufferList[offset+2] = *valueBytes
-	}
-
-	// Convert from an array of byte arrays to a single byte array
-	var entityKeyBuffer []byte
-	for i := 0; i < len(bufferList); i++ {
-		entityKeyBuffer = append(entityKeyBuffer, bufferList[i]...)
-	}
-
-	return &entityKeyBuffer, nil
-}
-
-func serializeCassandraValue(value interface{}, entityKeySerializationVersion int64) (*[]byte, types.ValueType_Enum, error) {
-	// TODO: Implement support for other types (at least the major types like ints, strings, bytes)
-	switch x := (value).(type) {
-	case *types.Value_StringVal:
-		valueString := []byte(x.StringVal)
-		return &valueString, types.ValueType_STRING, nil
-	case *types.Value_BytesVal:
-		return &x.BytesVal, types.ValueType_BYTES, nil
-	case *types.Value_Int32Val:
-		valueBuffer := make([]byte, 4)
-		binary.LittleEndian.PutUint32(valueBuffer, uint32(x.Int32Val))
-		return &valueBuffer, types.ValueType_INT32, nil
-	case *types.Value_Int64Val:
-		if entityKeySerializationVersion <= 1 {
-			//  We unfortunately have to use 32 bit here for backward compatibility :(
-			valueBuffer := make([]byte, 4)
-			binary.LittleEndian.PutUint32(valueBuffer, uint32(x.Int64Val))
-			return &valueBuffer, types.ValueType_INT64, nil
-		} else {
-			valueBuffer := make([]byte, 8)
-			binary.LittleEndian.PutUint64(valueBuffer, uint64(x.Int64Val))
-			return &valueBuffer, types.ValueType_INT64, nil
-		}
-	case nil:
-		return nil, types.ValueType_INVALID, fmt.Errorf("could not detect type for %v", x)
-	default:
-		return nil, types.ValueType_INVALID, fmt.Errorf("could not detect type for %v", x)
-	}
 }
 
 func (c *CassandraOnlineStore) Destruct() {
